+++ src/components/helpMenu/HelpMenu.jsx
@@ -1,4 +1,16 @@
-const HelpMenu = ({ title, width, position, isOpen, onRequestClose }) => {
+import { useHelpData } from "../../contexts/HelpDataContext";
+
+const HelpMenu = ({
+  title,
+  width,
+  position,
+  isOpen,
+  onRequestClose,
+
+
+}) => {
+  const {helpDataObject} = useHelpData()
+  const { introduction, gettingStarted, faqs, advancedFeatures } = helpDataObject;
   return (
     <div
       className={`fixed top-0 right-0 bg-black/[0.65] w-full h-screen z-40 transition-opacity duration-300 
@@ -8,14 +20,25 @@ const HelpMenu = ({ title, width, position, isOpen, onRequestClose }) => {
         className={`bg-white px-8 py-12 ${width} absolute ${position} h-screen shadow-lg transform transition-transform duration-700
         ${isOpen ? "translate-x-0" : "translate-x-full"} overflow-y-auto`}
       >
-        <h2 className="text-blue-300 text-lg font-bold pt-40 sm:pt-40 md:pt-40">{title}</h2>
+        <h2 className="text-blue-300 text-lg font-bold pt-40 sm:pt-40 md:pt-40">
+          {title}
+        </h2>
         <button
           onClick={onRequestClose}
           className="absolute top-4 right-4 pt-40 sm:pt-30 md:pt-45 text-gray-900 text-xl hover:text-gray-900 cursor-pointer"
         >
           x
         </button>
+        <div className="flex flex-col justify-center items-center w-full">
+          <label>Search Topics</label>
+          <input className="bg-neutral-300 w-5/6" type="text" />
+          <p>{introduction.title}</p>
+        </div>
+        <div>
+          <h1>gettingStarted</h1>
+        </div>
       </div>
+
     </div>
   );
 };
